refactor(cli,vpn): align Linux implementation with Windows
